[JT-68] refactor: 패키지 명 변경
